+++ src/Exception/ClassNotInstantiableException.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Exception;
+
+class ClassNotInstantiableException extends NotMappableValueException
+{
+    /**
+     * @param class-string $class
+     */
+    public function __construct(string $class)
+    {
+        parent::__construct(sprintf('Trying to instantiate class "%s", but this class is not instantiable.', $class));
+    }
+}